// FrontEnd-Service/src/presentation/components/forms/Screening/ScreeningEditForm.types.ts
import { GenreEnum } from "@infrastructure/apis/client/models/GenreEnum";
import { FormController } from "../FormController";
import { ScreenEnum } from "@infrastructure/apis/client/models/ScreenEnum";
import {
    UseFormHandleSubmit,
    UseFormRegister,
    FieldErrorsImpl,
    DeepRequired,
    UseFormWatch
} from "react-hook-form";
import { SelectChangeEvent } from "@mui/material";

export type ScreeningEditFormModel = {
    id: string;
    startTime: Date;
    movieId: string;
    hallId: string;
    screenType: ScreenEnum;
};

export type ScreeningEditFormState = {
    errors: FieldErrorsImpl<DeepRequired<ScreeningEditFormModel>>;
};

export type ScreeningEditFormActions = {
    register: UseFormRegister<ScreeningEditFormModel>;
    watch: UseFormWatch<ScreeningEditFormModel>;
    handleSubmit: UseFormHandleSubmit<ScreeningEditFormModel>;
    submit: (body: ScreeningEditFormModel) => void;
    radioScreenType: (event: React.ChangeEvent<HTMLInputElement>) => void;
    cancelOption: () => void;
    selectMovie: (value: SelectChangeEvent<string>) => void;
    selectHall: (value: SelectChangeEvent<string>) => void;
};
export type ScreeningEditFormComputed = {
    defaultValues: ScreeningEditFormModel,
    isSubmitting: boolean
};

export type ScreeningEditFormController = FormController<ScreeningEditFormState, ScreeningEditFormActions, ScreeningEditFormComputed>;